Calculadora de Estadísticas y Búsqueda en un Arreglo: Se modifica y se avanza en el código.

+++ Tareas/Tarea1/src/Programa2/Estadistica_Arreglo.hpp
@@ -43,14 +43,14 @@ public:
     Estadisticas(double* numeros, int cantidadNumeros);
 
     // Métodos para calcular las estadísticas y encontrar número
-    int calcularSuma();
+    double calcularSuma();
     double calcularPromedio();
-    int encontrarMaximo();
-    int encontrarMinimo();
+    double encontrarMaximo();
+    double encontrarMinimo();
     bool buscarNumero(int numero);
 
 private:
-    int* numeros;  // Puntero al arreglo de números
+    double* numeros;  // Puntero al arreglo de números
     int cantidadNumeros;   // Tamaño del arreglo
 };
